+++ packages/runtime-core/src/componentOptions.ts
@@ -85,6 +85,7 @@ import {
   type ComponentTypeEmits,
   normalizePropsOrEmits,
 } from './apiSetupHelpers'
+import { markAsyncBoundary } from './helpers/useId'
 
 /**
  * Interface for declaring custom options.
@@ -771,6 +772,10 @@ export function applyOptions(instance: ComponentInternalInstance) {
     isCompatEnabled(DeprecationTypes.FILTERS, instance)
   ) {
     instance.filters = filters
+  }
+
+  if (__SSR__ && serverPrefetch) {
+    markAsyncBoundary(instance)
   }
 }