+++ scripts/release.js
@@ -1,13 +1,13 @@
 // @ts-check
-import minimist from 'minimist'
 import fs from 'node:fs'
 import path from 'node:path'
 import pico from 'picocolors'
 import semver from 'semver'
 import enquirer from 'enquirer'
-import { execa } from 'execa'
 import { createRequire } from 'node:module'
 import { fileURLToPath } from 'node:url'
+import { exec } from './utils.js'
+import { parseArgs } from 'node:util'
 
 /**
  * @typedef {{
@@ -23,12 +23,37 @@ let versionUpdated = false
 const { prompt } = enquirer
 const currentVersion = createRequire(import.meta.url)('../package.json').version
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
-const args = minimist(process.argv.slice(2), {
-  alias: {
-    skipBuild: 'skip-build',
-    skipTests: 'skip-tests',
-    skipGit: 'skip-git',
-    skipPrompts: 'skip-prompts',
+
+const { values: args, positionals } = parseArgs({
+  allowPositionals: true,
+  options: {
+    preid: {
+      type: 'string',
+    },
+    dry: {
+      type: 'boolean',
+    },
+    tag: {
+      type: 'string',
+    },
+    canary: {
+      type: 'boolean',
+    },
+    skipBuild: {
+      type: 'boolean',
+    },
+    skipTests: {
+      type: 'boolean',
+    },
+    skipGit: {
+      type: 'boolean',
+    },
+    skipPrompts: {
+      type: 'boolean',
+    },
+    vapor: {
+      type: 'boolean',
+    },
   },
 })
 
@@ -111,16 +136,16 @@ const versionIncrements = [
 ]
 
 const inc = (/** @type {import('semver').ReleaseType} */ i) =>
-  semver.inc(currentVersion, i, preId)
+  semver.inc(currentVersion, i, typeof preId === 'string' ? preId : undefined)
 const run = async (
   /** @type {string} */ bin,
   /** @type {ReadonlyArray<string>} */ args,
-  /** @type {import('execa').Options} */ opts = {},
-) => execa(bin, args, { stdio: 'inherit', ...opts })
+  /** @type {import('node:child_process').SpawnOptions} */ opts = {},
+) => exec(bin, args, { stdio: 'inherit', ...opts })
 const dryRun = async (
   /** @type {string} */ bin,
   /** @type {ReadonlyArray<string>} */ args,
-  /** @type {import('execa').Options} */ opts = {},
+  /** @type {import('node:child_process').SpawnOptions} */ opts = {},
 ) => console.log(pico.blue(`[dryrun] ${bin} ${args.join(' ')}`), opts)
 const runIfNotDry = isDryRun ? dryRun : run
 const getPkgRoot = (/** @type {string} */ pkg) =>
@@ -134,7 +159,7 @@ async function main() {
     console.log(`${pico.green(`✓`)} commit is up-to-date with remote.\n`)
   }
 
-  let targetVersion = args._[0]
+  let targetVersion = positionals[0]
 
   if (isCanary || isVapor) {
     const major = semver.major(currentVersion)
@@ -425,12 +450,12 @@ async function isInSyncWithRemote() {
  */
 async function getSha(short) {
   return (
-    await execa('git', ['rev-parse', ...(short ? ['--short'] : []), 'HEAD'])
+    await exec('git', ['rev-parse', ...(short ? ['--short'] : []), 'HEAD'])
   ).stdout
 }
 
 async function getBranch() {
-  return (await execa('git', ['rev-parse', '--abbrev-ref', 'HEAD'])).stdout
+  return (await exec('git', ['rev-parse', '--abbrev-ref', 'HEAD'])).stdout
 }
 
 /**